feat & fix: migration and data seeder

+++ src/database/migrations/2024_11_15_001442_create_categories_table.php
@@ -15,7 +15,7 @@ return new class extends Migration
             $table->id();
             $table->string('name', 100)->nullable();
             $table->text('description')->nullable();
-            $table->unsignedInteger('created_by')->nullable();
+            $table->unsignedBigInteger('created_by')->nullable();
             $table->foreign('created_by')->references('id')->on('admins')->onDelete('restrict');
             $table->timestamps();
         });
